+++ cansend.c
@@ -156,7 +156,7 @@ int main(int argc, char **argv)
 		}
 		mtu = ifr.ifr_mtu;
 
-		if (mtu == CANFD_MTU) {
+		if (mtu == (int)CANFD_MTU) {
 			/* interface is ok - try to switch the socket into CAN FD mode */
 			if (setsockopt(s, SOL_CAN_RAW, CAN_RAW_FD_FRAMES,
 				       &enable_canfx, sizeof(enable_canfx))){
@@ -165,7 +165,7 @@ int main(int argc, char **argv)
 			}
 		}
 
-		if (mtu >= CANXL_MIN_MTU) {
+		if (mtu >= (int)CANXL_MIN_MTU) {
 			/* interface is ok - try to switch the socket into CAN XL mode */
 			if (setsockopt(s, SOL_CAN_RAW, CAN_RAW_XL_FRAMES,
 				       &enable_canfx, sizeof(enable_canfx))){
